qa: fixed two minor issues

+++ docker/startup-dev.sh
@@ -2,8 +2,8 @@ if [ ! -d "node_modules" ]; then
     npm install
 fi
 
-# if [ ! -f ".env.local" ]; then
+if [ ! -f ".env.local" ]; then
     cp docker/docker-env .env.local
-# fi
+fi
 
 npm run dev
